backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
httpx==0.27.2
groq==0.4.2
google-generativeai==0.5.0
python-dotenv==1.0.0
pydantic==2.11.4
